requirements.txt
playwright==1.43.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.0.0
schedule==1.2.0
